create get_policy
