+++ onedal/dal.cpp
@@ -23,8 +23,6 @@ namespace oneapi::dal::python {
 
 /* common */
 #ifdef ONEDAL_DATA_PARALLEL_SPMD
-ONEDAL_PY_INIT_MODULE(spmd_policy);
-
 /* algorithms */
 ONEDAL_PY_INIT_MODULE(covariance);
 ONEDAL_PY_INIT_MODULE(dbscan);
@@ -84,9 +82,7 @@ ONEDAL_PY_INIT_MODULE(finiteness_checker);
 #endif // ONEDAL_DATA_PARALLEL_SPMD
 
 #ifdef ONEDAL_DATA_PARALLEL_SPMD
-PYBIND11_MODULE(_onedal_py_spmd_dpc, m) {
-    init_spmd_policy(m);
-
+PYBIND11_MODULE(_onedal_py_spmd_dpc, m) {   
     init_covariance(m);
     init_dbscan(m);
     init_decomposition(m);
